requirements.txt
absl-py==0.4.1
astor==0.7.1
Augmentor==0.2.3
backcall==0.1.0
bleach==1.5.0
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
dill==0.2.8.2
Django==2.1.1
dopamine-rl==2.0.1
entrypoints==0.3
enum34==1.1.6
flake8==3.7.7
Flask==1.0.2
future==0.17.1
gast==0.2.0
gevent==1.4.0
gin-config==0.1.4
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.5.8
greenlet==0.4.15
grpcio==1.15.0
gunicorn==19.9.0
gym==0.12.0
h5py==2.8.0
html5lib==0.9999999
httplib2==0.12.1
idna==2.8
image==1.5.25
imgaug==0.2.6
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jax==0.1.21
jaxlib==0.1.11
jedi==0.12.1
Jinja2==2.10
Keras==2.0.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kfac==0.1.4
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mesh-tensorflow==0.0.5
mock==2.0.0
mpmath==1.1.0
multiprocess==0.70.6.1
munch==2.3.2
natsort==5.4.1
networkx==2.2
numpy==1.15.1
oauth2client==4.1.3
opencv-python==3.4.4.19
opt-einsum==2.3.2
parso==0.3.1
pathos==0.2.2.1
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pox==0.2.4
ppft==1.6.4.8
promise==2.2.1
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.7
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-image==0.14.1
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sympy==1.3
tensor2tensor==1.13.0
tensorboard==1.13.1
tensorflow-datasets==1.0.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-metadata==0.13.0
tensorflow-probability==0.6.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tf-agents==0.2.0rc2
toolz==0.9.0
tqdm==4.28.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
